Move ZoomableScrollPane
