+++ core/src/com/unciv/ui/screens/mapeditorscreen/EditorMapHolder.kt
@@ -15,7 +15,7 @@ import com.unciv.ui.components.tilegroups.TileGroupMap
 import com.unciv.ui.components.tilegroups.TileGroup
 import com.unciv.ui.components.tilegroups.TileSetStrings
 import com.unciv.ui.screens.basescreen.BaseScreen
-import com.unciv.ui.components.ZoomableScrollPane
+import com.unciv.ui.components.widgets.ZoomableScrollPane
 import com.unciv.ui.components.input.onClick
 import com.unciv.ui.screens.basescreen.UncivStage
