+++ core/src/com/unciv/ui/screens/worldscreen/ZoomButtonPair.kt
@@ -5,7 +5,7 @@ import com.badlogic.gdx.scenes.scene2d.ui.Table
 import com.badlogic.gdx.scenes.scene2d.ui.TextButton
 import com.badlogic.gdx.utils.Align
 import com.unciv.ui.screens.basescreen.BaseScreen
-import com.unciv.ui.components.ZoomableScrollPane
+import com.unciv.ui.components.widgets.ZoomableScrollPane
 import com.unciv.ui.components.input.onClick
 import com.unciv.ui.components.extensions.setFontSize
